# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTodoRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      dueDate:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.Priority'
        enum:
        - low
        - medium
        - high
    required:
    - description
    - priority
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  models.Priority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
...
  models.Todo:
    properties:
      completed:
        type: boolean
      completedAt:
        type: string
      createdAt:
        type: string
      description:
        maxLength: 500
        minLength: 1
        type: string
      dueDate:
        type: string
      id:
        type: string
      listId:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.Priority'
        enum:
        - low
        - medium
        - high
      updatedAt:
        type: string
    required:
    - description
    - priority
    type: object
...
  models.UpdateTodoRequest:
    properties:
      completed:
        type: boolean
      description:
        maxLength: 500
        minLength: 1
        type: string
      dueDate:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.Priority'
        enum:
        - low
        - medium
        - high
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@todolist-api.com
    name: API Support
  description: A RESTful API for managing todo lists and tasks with user authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TodoList API
  version: "1.0"
paths:
...
  /lists/{listId}/todos:
    get:
      description: Get all todos in a specific list with optional filtering and sorting
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Filter by priority (low, medium, high)
        in: query
        name: priority
        type: string
      - description: Filter by completion status
        in: query
        name: completed
        type: boolean
      - description: Sort by field (due_date, priority, created_at)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get todos in a list
      tags:
      - Todos
    post:
      consumes:
      - application/json
      description: Create a new todo item in a specific list
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Todo details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new todo
      tags:
      - Todos
  /lists/{listId}/todos/{todoId}:
    delete:
      description: Delete a specific todo item
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Todo ID (UUID)
        in: path
        name: todoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a todo
      tags:
      - Todos
    get:
      description: Get a specific todo item by ID
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Todo ID (UUID)
        in: path
        name: todoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a specific todo
      tags:
      - Todos
    put:
      consumes:
      - application/json
      description: Update a specific todo item
      parameters:
      - description: List ID (UUID)
        in: path
        name: listId
        required: true
        type: string
      - description: Todo ID (UUID)
        in: path
        name: todoId
        required: true
        type: string
      - description: Updated todo details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a todo
      tags:
      - Todos
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...